Iterate on mermaid refactor
